fix : Change data type for VOs
 - 메인페이지 그래프 출력오류 수정
 - VO내 amount 변수를 int -> long 으로 변경

+++ src/main/java/com/modernhome/domain/OutOrderResultVO.java
@@ -4,7 +4,7 @@ public class OutOrderResultVO {
 	
 	private String month;
 	private Integer monthlyOrderCount;
-	private Integer monthlyOrderAmount;
+	private long monthlyOrderAmount;
 	
 	
 	public Integer getMonthlyOrderCount() {
@@ -13,10 +13,10 @@ public class OutOrderResultVO {
 	public void setMonthlyOrderCount(Integer monthlyOrderCount) {
 		this.monthlyOrderCount = monthlyOrderCount;
 	}
-	public Integer getMonthlyOrderAmount() {
+	public long getMonthlyOrderAmount() {
 		return monthlyOrderAmount;
 	}
-	public void setMonthlyOrderAmount(Integer monthlyOrderAmount) {
+	public void setMonthlyOrderAmount(long monthlyOrderAmount) {
 		this.monthlyOrderAmount = monthlyOrderAmount;
 	}
